requirements.txt
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pytest>=8.0.0
ruff>=0.4.0
